# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mapa de Emergencia y Rescate — API",
    "version": "1.0.0",
    "description": "API pública/admin del mapa de emergencia. Documentación generada automáticamente desde los bloques @swagger de cada route."
  },
...
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "Hospital": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "externalId": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "facilityType": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "municipality": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "level": {
            "type": "string",
            "nullable": true
          },
          "priorityZone": {
            "type": "string",
            "enum": [
              "P0",
              "P1",
              "P2",
              "P3"
            ]
          },
          "isPriority": {
            "type": "boolean"
          },
          "activePatients": {
            "type": "integer"
          },
          "totalPatients": {
            "type": "integer"
          },
          "createdAt": {
            "type": "integer"
          },
          "supplySummary": {
            "nullable": true,
            "$ref": "#/components/schemas/HospitalSupplySummary"
          }
        }
      },
...
      "HospitalSupplyFreshness": {
        "type": "object",
        "properties": {
          "lastUpdatedAt": {
            "type": "integer",
            "description": "epoch-ms"
          },
          "lastConfirmedAt": {
            "type": "integer",
            "description": "epoch-ms"
          },
          "staleAfterHours": {
            "type": "integer"
          },
          "isStale": {
            "type": "boolean"
          },
          "updatedAgo": {
            "type": "string"
          },
          "confirmedAgo": {
            "type": "string"
          }
        }
      },
      "HospitalSupplyCategory": {
        "type": "string",
        "enum": [
          "medications",
          "iv_fluids",
          "medical_supplies",
          "soft_foods",
          "water",
          "beds_capacity",
          "lab_diagnostics",
          "transport",
          "other"
        ]
      },
      "HospitalSupplySemaphore": {
        "type": "string",
        "enum": [
          "green",
          "yellow",
          "red",
          "unknown"
        ]
      },
      "HospitalSupplyUrgencySemaphore": {
        "type": "string",
        "enum": [
          "yellow",
          "red",
          "unknown"
        ],
        "description": "Urgencia permitida para necesidades activas; verde se reserva para semáforos sin necesidad crítica."
      },
      "HospitalSupplyCategoryStatus": {
        "type": "object",
        "properties": {
          "category": {
            "$ref": "#/components/schemas/HospitalSupplyCategory"
          },
          "status": {
            "$ref": "#/components/schemas/HospitalSupplySemaphore"
          },
          "label": {
            "type": "string"
          },
          "publicNote": {
            "type": "string"
          },
          "freshness": {
            "$ref": "#/components/schemas/HospitalSupplyFreshness"
          }
        }
      },
...
      "HospitalSupplyNeed": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "hospitalId": {
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/HospitalSupplyCategory"
          },
          "categoryLabel": {
            "type": "string"
          },
          "itemType": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "nullable": true
          },
          "unit": {
            "type": "string"
          },
          "urgency": {
            "$ref": "#/components/schemas/HospitalSupplyUrgencySemaphore"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "partially_covered",
              "covered",
              "cancelled",
              "needs_verification"
            ]
          },
          "publicNote": {
            "type": "string"
          },
          "lastConfirmedAt": {
            "type": "integer"
          },
          "updatedAt": {
            "type": "integer"
          },
          "updatedAgo": {
            "type": "string"
          }
        }
      },
...
      "HospitalSupplySummary": {
        "type": "object",
        "properties": {
          "statuses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HospitalSupplyCategoryStatus"
            }
          },
          "activeNeeds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HospitalSupplyNeed"
            }
          },
          "counts": {
            "type": "object",
            "properties": {
              "red": {
                "type": "integer"
              },
              "yellow": {
                "type": "integer"
              },
              "stale": {
                "type": "integer"
              },
              "activeNeeds": {
                "type": "integer"
              }
            }
          },
          "worstStatus": {
            "$ref": "#/components/schemas/HospitalSupplySemaphore"
          },
          "lastConfirmedAt": {
            "type": "integer",
            "nullable": true
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/hospitals": {
      "get": {
        "tags": [
          "hospitals"
        ],
        "summary": "Lista hospitales con filtros opcionales y estados",
        "parameters": [
          {
            "in": "query",
            "name": "include",
            "schema": {
              "type": "string",
              "enum": [
                "states"
              ]
            },
            "description": "Si es \"states\", incluye la lista de estados"
          },
          {
            "in": "query",
            "name": "zone",
            "schema": {
              "type": "string",
              "enum": [
                "P0",
                "P1",
                "P2",
                "P3"
              ]
            },
            "description": "Zona de prioridad"
          },
          {
            "in": "query",
            "name": "state",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            },
            "description": "Texto de búsqueda"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 500
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hospitales y, opcionalmente, estados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "hospitals": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Hospital"
                      }
                    },
                    "states": {
                      "type": "array",
                      "nullable": true,
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "hospitals"
        ],
        "summary": "Registra un nuevo hospital",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "state"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "facilityType": {
                    "type": "string"
                  },
                  "state": {
                    "type": "string"
                  },
                  "municipality": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  },
                  "level": {
                    "type": "string"
                  },
                  "priorityZone": {
                    "type": "string",
                    "enum": [
                      "P0",
                      "P1",
                      "P2",
                      "P3"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Hospital creado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "hospital": {
                      "$ref": "#/components/schemas/Hospital"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos (nombre o estado faltante o nombre muy largo)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Demasiadas peticiones (rate limit)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "No se pudo guardar el hospital",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}